+++ test/Injector/ApplicationConfigInjectorTest.php
@@ -17,6 +17,8 @@ class ApplicationConfigInjectorTest extends AbstractInjectorTestCase
     protected $injectorTypesAllowed = [
         ApplicationConfigInjector::TYPE_COMPONENT,
         ApplicationConfigInjector::TYPE_MODULE,
+        ApplicationConfigInjector::TYPE_DEPENDENCY,
+        ApplicationConfigInjector::TYPE_BEFORE_APPLICATION,
     ];
 
     public function allowedTypes()
@@ -25,6 +27,8 @@ class ApplicationConfigInjectorTest extends AbstractInjectorTestCase
             'config-provider' => [ApplicationConfigInjector::TYPE_CONFIG_PROVIDER, false],
             'component'       => [ApplicationConfigInjector::TYPE_COMPONENT, true],
             'module'          => [ApplicationConfigInjector::TYPE_MODULE, true],
+            'dependency'      => [ApplicationConfigInjector::TYPE_DEPENDENCY, true],
+            'before-application-modules' => [ApplicationConfigInjector::TYPE_BEFORE_APPLICATION, true],
         ];
     }