+++ test/Injector/DevelopmentConfigInjectorTest.php
@@ -17,6 +17,8 @@ class DevelopmentConfigInjectorTest extends AbstractInjectorTestCase
     protected $injectorTypesAllowed = [
         DevelopmentConfigInjector::TYPE_COMPONENT,
         DevelopmentConfigInjector::TYPE_MODULE,
+        DevelopmentConfigInjector::TYPE_DEPENDENCY,
+        DevelopmentConfigInjector::TYPE_BEFORE_APPLICATION,
     ];
 
     public function allowedTypes()
@@ -25,6 +27,8 @@ class DevelopmentConfigInjectorTest extends AbstractInjectorTestCase
             'config-provider' => [DevelopmentConfigInjector::TYPE_CONFIG_PROVIDER, false],
             'component'       => [DevelopmentConfigInjector::TYPE_COMPONENT, true],
             'module'          => [DevelopmentConfigInjector::TYPE_MODULE, true],
+            'dependency'      => [DevelopmentConfigInjector::TYPE_DEPENDENCY, true],
+            'before-application-modules' => [DevelopmentConfigInjector::TYPE_BEFORE_APPLICATION, true],
         ];
     }